+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -4,21 +4,20 @@ package ru.yandex.practicum.filmorate.model;
 import lombok.Builder;
 import lombok.Data;
 
-import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
 @Data
 @Builder
 public class Review {
-    private int reviewId;
+    private long reviewId;
     @NotBlank
     private String content;
     @NotNull
     private Boolean isPositive;
     @NotNull
-    private Integer userId;
+    private long userId;
     @NotNull
-    private Integer filmId;
-    private int useful;
+    private long filmId;
+    private long useful;
 }
